Validating uuid with regex and moving /course/books to /course to match API doc
