+++ lib/helpers/backend/valid-uuid.ts
@@ -1,0 +1,4 @@
+export function isValidUUID(id : string): boolean {
+    const regex: RegExp = /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/
+    return regex.test(id)
+}
